separate FlashParam from ConfigParam
