--- src/ConfigParam.cpp
@@ -1,127 +1,0 @@
-/*-----------------------------------------------------------/
-/ ConfigParam.h
-/------------------------------------------------------------/
-/ Copyright (c) 2024, Elehobica
-/ Released under the BSD-2-Clause
-/ refer to https://opensource.org/licenses/BSD-2-Clause
-/-----------------------------------------------------------*/
-
-#include "ConfigParam.h"
-
-namespace ConfigParamNs {
-
-//=================================
-// Implementation of Parameter class
-//=================================
-template <class T>
-Parameter<T>::Parameter(const ParamId_t& id, const char* name, const uint32_t& flashAddr, const valueType& defaultValue, size_t& size)
-    : id(id), name(name), flashAddr(flashAddr), defaultValue(defaultValue), size(size)
-{
-    Params::instance().add(id, this);
-}
-
-template <class T>
-Parameter<T>::Parameter(const ParamId_t& id, const char* name, const uint32_t& flashAddr, const valueType& defaultValue)
-    : id(id), name(name), flashAddr(flashAddr), defaultValue(defaultValue), size(sizeof(T))
-{
-    Params::instance().add(id, this);
-}
-
-//=================================
-// Implementation of Params class
-//=================================
-Params& Params::instance()
-{
-    static Params instance; // Singleton
-    return instance;
-}
-
-void Params::printInfo() const
-{
-    printf("=== ConfigParam ===\n");
-    for (const auto& [key, item] : paramMap) {
-        std::visit(PrintInfoVisitor{}, item);
-    }
-}
-
-void Params::loadDefault()
-{
-    for (auto& [key, item] : paramMap) {
-        std::visit([](auto&& param) {
-            param->loadDefault();
-        }, item);
-    }
-}
-
-void Params::loadFromFlash()
-{
-    for (auto& [key, item] : paramMap) {
-        std::visit(ReadFromFlashVisitor{}, item);
-    }
-}
-
-void Params::storeToFlash() const
-{
-    for (const auto& [key, item] : Params::instance().paramMap) {
-        std::visit(WriteReserveVisitor{}, item);
-    }
-    UserFlash& userFlash = UserFlash::instance();
-    userFlash.program();
-}
-
-//=================================
-// Implementation of ConfigParamClass class
-//=================================
-ConfigParamClass& ConfigParamClass::instance()
-{
-    static ConfigParamClass instance; // Singleton
-    return instance;
-}
-
-void ConfigParamClass::printInfo() const
-{
-    Params::instance().printInfo();
-}
-
-void ConfigParamClass::loadDefault()
-{
-    Params::instance().loadDefault();
-}
-
-uint32_t ConfigParamClass::getBootCountFromFlash()
-{
-    auto& param = P_CFG_BOOT_COUNT;
-    ReadFromFlashVisitor visitor;
-    visitor(&param);
-    return param.get();
-}
-
-void ConfigParamClass::incBootCount()
-{
-    auto& param = P_CFG_BOOT_COUNT;
-    param.set(param.get() + 1);
-}
-
-void ConfigParamClass::initialize()
-{
-    // load default
-    loadDefault();
-    // don't load from Flash if flash is blank
-    if (getBootCountFromFlash() == 0xffffffffUL) { return; }
-
-    // load from flash and get format revision
-    uint32_t formatRevExpected = P_CFG_FORMAT_REV.get();
-    Params::instance().loadFromFlash();
-    uint32_t formatRev = P_CFG_FORMAT_REV.get();
-
-    // Force to reset to default due to format revision changed (parameter/address) to avoid mulfunction
-    if (formatRevExpected != formatRev) {
-        loadDefault();
-    }
-}
-
-void ConfigParamClass::finalize() const
-{
-    Params::instance().storeToFlash();
-}
-}
