+++ src/ConfigParam.h
@@ -8,15 +8,9 @@
 
 #pragma once
 
-#include <map>
-#include <string>
-#include <cinttypes>  // this must be located at later than <string>
-#include <variant>
+#include "FlashParam.h"
 
-#include "UserFlash.h"
-
-
-namespace ConfigParamNs {
+namespace FlashParamNs {
 typedef enum {
     CFG_BOOT_COUNT = 0,
     CFG_FORMAT_REV,
@@ -53,127 +47,32 @@ typedef enum {
     CFG_MENU_IDX_PLAY_RANDOM_DIR_DEPTH,
 } ParamId_t;
 
-//=================================
-// Interface of Parameter class
-//=================================
-template <class T>
-class Parameter {
-    using valueType = T;
-public:
-    void set(valueType value_) { value = value_; }
-    valueType get() const { return value; }
-private:
-    Parameter(const ParamId_t& id, const char* name, const uint32_t& flashAddr, const valueType& defaultValue, size_t& size);
-    Parameter(const ParamId_t& id, const char* name, const uint32_t& flashAddr, const valueType& defaultValue);
-    Parameter(const Parameter&) = delete;
-    Parameter& operator=(const Parameter&) = delete;  // don't permit copy
-    void loadDefault() { value = defaultValue; }
-    const ParamId_t id;
-    const char* name;
-    const uint32_t flashAddr;
-    const valueType defaultValue;
-    const size_t size;
-    valueType value = defaultValue;
-    friend class Params;
-    friend class ConfigParamClass;
-    friend class ReadFromFlashVisitor;
-    friend class WriteReserveVisitor;
-    friend class PrintInfoVisitor;
-};
-
-using variant_t = std::variant<
-    Parameter<bool>*,
-    Parameter<uint8_t>*, Parameter<uint16_t>*, Parameter<uint32_t>*, Parameter<uint64_t>*,
-    Parameter<int8_t>*, Parameter<int16_t>*, Parameter<int32_t>*, Parameter<int64_t>*,
-    Parameter<char*>*>;
-
-//=================================
-// Interface of Visitors
-//=================================
-struct FlashVisitor {
-    UserFlash& userFlash = UserFlash::instance();
-};
-
-struct ReadFromFlashVisitor : FlashVisitor {
-    template <typename T>
-    void operator()(const T& param) const {
-        userFlash.read(param->flashAddr, param->size, param->value);
+struct ConfigParam : ConfigParamBase {
+    static ConfigParam& instance() {  // Singleton
+        static ConfigParam instance;
+        return instance;
     }
-};
-
-struct WriteReserveVisitor : FlashVisitor {
-    template <typename T>
-    void operator()(const T& param) const {
-        userFlash.writeReserve(param->flashAddr, param->size, param->value);
+    uint32_t getBootCountFromFlash() {
+        auto& param = P_CFG_BOOT_COUNT;
+        ReadFromFlashVisitor visitor;
+        visitor(&param);
+        return param.get();
     }
-};
-
-struct PrintInfoVisitor {
-    template <typename T>
-    void operator()(T& param) const {
-        const variant_t item = param;
-        const auto& format = printFormat.at(item.index());
-        printf(format.c_str(), param->flashAddr, param->name, param->value, param->value);
+    void incBootCount() {
+        auto& param = P_CFG_BOOT_COUNT;
+        param.set(param.get() + 1);
     }
-    const std::array<std::string, 10> printFormat = {  // this must be matched with variant order due to being referred by index()
-        "0x%04x %s: %" PRIi32 "d (0x%" PRIx32 ")\n",  // bool
-        "0x%04x %s: %" PRIu8 "d (0x%" PRIx8 ")\n",    // uint8_t
-        "0x%04x %s: %" PRIu16 "d (0x%" PRIx16 ")\n",  // uint16_t
-        "0x%04x %s: %" PRIu32 "d (0x%" PRIx32 ")\n",  // uint32_t
-        "0x%04x %s: %" PRIu64 "d (0x%" PRIx64 ")\n",  // uint64_t
-        "0x%04x %s: %" PRIi8 "d (0x%" PRIx8 ")\n",    // int8_t
-        "0x%04x %s: %" PRIi16 "d (0x%" PRIx16 ")\n",  // int16_t
-        "0x%04x %s: %" PRIi32 "d (0x%" PRIx32 ")\n",  // int32_t
-        "0x%04x %s: %" PRIi64 "d (0x%" PRIx64 ")\n",  // int64_t
-        "0x%04x %s: %s\n",                            // char*
-    };
-};
-
-//=================================
-// Interface of Params class
-//=================================
-class Params
-{
-// all private except for friend classes
-    static Params& instance(); // Singleton
-    void printInfo() const;
-    void loadDefault();
-    void loadFromFlash();
-    void storeToFlash() const;
-    template <typename T>
-    void add(const ParamId_t& id, T* paramPtr) { paramMap[id] = paramPtr; }
-    template <typename T>
-    T& getParam(const ParamId_t& id) {
-        auto& item = paramMap.at(id);
-        auto& paramPtr = std::get<T*>(item);
-        return *paramPtr;
+    void initialize() {
+        loadDefault();
+        // don't load from Flash if flash is blank
+        if (getBootCountFromFlash() == 0xffffffffUL) { return; }
+        // load from flash and get format revision
+        uint32_t formatRevExpected = P_CFG_FORMAT_REV.get();
+        loadFromFlash();
+        uint32_t formatRev = P_CFG_FORMAT_REV.get();
+        // Force to reset to default due to format revision changed (parameter/address) to avoid mulfunction
+        if (formatRevExpected != formatRev) { loadDefault(); }
     }
-    std::map<const ParamId_t, variant_t> paramMap;
-    template<typename> friend class Parameter;  // for all Parameter<> classes
-    friend class ConfigParamClass;
-};
-
-//=================================
-// Interface of ConfigParamClass class
-//=================================
-class ConfigParamClass
-{
-public:
-    static ConfigParamClass& instance(); // Singleton
-    void printInfo() const;
-    void initialize();
-    void finalize() const;
-    void incBootCount();
-    // accessor by Parameter<> instance on template T = Parameter<>
-    template <typename T>
-    decltype(auto) getValue(const T& param) const { return param.get(); }
-    template <typename T>
-    void setValue(T& param, const typename T::valueType value) { param.set(value); }
-    // accessor by ParamId_t on template T = primitive type
-    template <typename T>
-    decltype(auto) getValue(const ParamId_t& id) const { return _getValue<Parameter<T>>(id); }
-    template <typename T>
-    void setValue(const ParamId_t& id, const T& value) { _setValue<Parameter<T>>(id, value); }
 
     // Parameter<T>     inst                                         id                                          name                                          addr   default
     Parameter<uint32_t> P_CFG_BOOT_COUNT                            {CFG_BOOT_COUNT,                             "CFG_BOOT_COUNT",                             0x000, 10};
@@ -210,27 +109,8 @@ public:
     Parameter<uint32_t> P_CFG_MENU_IDX_PLAY_TIME_TO_NEXT_PLAY       {CFG_MENU_IDX_PLAY_TIME_TO_NEXT_PLAY,        "CFG_MENU_IDX_PLAY_TIME_TO_NEXT_PLAY",        0x0a4, 2};
     Parameter<uint32_t> P_CFG_MENU_IDX_PLAY_NEXT_PLAY_ALBUM         {CFG_MENU_IDX_PLAY_NEXT_PLAY_ALBUM,          "CFG_MENU_IDX_PLAY_NEXT_PLAY_ALBUM",          0x0a8, 1};
     Parameter<uint32_t> P_CFG_MENU_IDX_PLAY_RANDOM_DIR_DEPTH        {CFG_MENU_IDX_PLAY_RANDOM_DIR_DEPTH,         "CFG_MENU_IDX_PLAY_RANDOM_DIR_DEPTH",         0x0ac, 1};
-
-private:
-    ConfigParamClass() = default;
-    ~ConfigParamClass() = default;
-    ConfigParamClass(const ConfigParamClass&) = delete;
-    ConfigParamClass& operator=(const ConfigParamClass&) = delete;
-    void loadDefault();
-    uint32_t getBootCountFromFlash();
-    // accessor by ParamId_t on template T = Patameter<>
-    template <typename T>
-    void _setValue(const ParamId_t& id, const typename T::valueType& value) {
-        auto& param = Params::instance().getParam<T>(id);
-        return param.set(value);
-    }
-    template <typename T>
-    decltype(auto) _getValue(const ParamId_t& id) const {
-        const auto& param = Params::instance().getParam<T>(id);
-        return param.get();
-    }
 };
 }
 
-// alias to ConfigParamNs::ConfigParamClass
-using ConfigParam = ConfigParamNs::ConfigParamClass;
+// alias to FlashParamNs::ConfigParam
+using ConfigParam = FlashParamNs::ConfigParam;
